src/backend_django/api/urls.py
"""
project_landing — API Configuration (Django Ninja).

Versioned API routers. Bot and external clients use these endpoints.
Docs: /api/docs
"""

from ninja import NinjaAPI, Router

api = NinjaAPI(
    title="project_landing API",
    version="1.0.0",
)

# ── v1 ──
v1 = Router()


@v1.get("/health")
def health(request):
    """Health check endpoint."""
    return {"status": "ok"}


api.add_router("/v1/", v1)
