pyproject.toml
[project]
name = "project_landing"
version = "0.1.0"
description = ""
authors = [
    {name = "CodexDLC",email = "primecodex@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"

# Общие зависимости — всегда включены
dependencies = [
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "django-modeltranslation (>=0.19.19,<0.20.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "holidays (>=0.63)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "aiosmtplib (>=3.0.1,<4.0.0)",
    "django-unfold (>=0.79.0,<0.80.0)",
]

[tool.poetry]
packages = [
    { include = "src" },
    { include = "tools" }
]

[tool.poetry.scripts]
check = "tools.dev.check:main"

# === Группы зависимостей ===

# Технические зависимости, общие для разных сервисов (бэкенд, бот, воркер)
[tool.poetry.group.shared.dependencies]
arq = ">=0.27.0,<0.28.0"

[tool.poetry.group.django.dependencies]
django = ">=5.1,<6.0"
python-dotenv = ">=1.0.0,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
dj-database-url = ">=2.3.0,<3.0.0"
gunicorn = ">=23.0.0,<24.0.0"
django-ninja = ">=1.3.0,<2.0.0"
django-redis = ">=5.4.0,<6.0.0"
django-prometheus = ">=2.3.1,<3.0.0"
django-ratelimit = ">=4.1.0,<5.0.0"

[tool.poetry.group.bot.dependencies]
aiogram = ">=3.24.0,<4.0.0"
aiogram-i18n = ">=1.4.0,<2.0.0"
"fluent.runtime" = ">=0.4.0,<0.5.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
mypy = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-django = "*"
pre-commit = "*"
qrcode = ">=7.0,<8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# === Ruff ===
[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"**/selectors/__init__.py" = ["A005"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# === Mypy ===
[tool.mypy]
python_version = "3.13"
strict = false
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = false
disable_error_code = ["call-arg"]

# === Pytest ===
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = [".", "src/backend_django"]
DJANGO_SETTINGS_MODULE = "core.settings.test"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
